+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased]
 - Added support for latest Leo and Aleo instructions grammar
+- Fixed broken variable and struct component completion
 
 ## [0.6.2]
 - Fixed highlighting of identifiers in Aleo instructions